Split up all the classes into different files/add directories
