+++ lib/raw_net_capture/raw_net_capture.rb
@@ -1,0 +1,24 @@
+# encoding: utf-8
+
+class RawNetCapture < StringIO
+  attr_reader :raw_traffic
+
+  def initialize
+    super
+    reset
+  end
+
+  def received(data)
+    @raw_traffic << [:received, data]
+  end
+
+  def sent(data)
+    @raw_traffic << [:sent, data]
+  end
+
+  private
+
+  def reset
+    @raw_traffic = []
+  end
+end
